+++ next.config.ts
@@ -1,9 +1,0 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  /* config options here */
-  output: "standalone",
-  allowedDevOrigins: ["127.0.0.1"],
-};
-
-export default nextConfig;